# Faster json parser for embeddings