--- a/cmd/cody-gateway/internal/httpapi/embeddings/BUILD.bazel
+++ b/cmd/cody-gateway/internal/httpapi/embeddings/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "//internal/httpcli",
         "//internal/trace",
         "//lib/errors",
+        "@com_github_go_json_experiment_json//:go_default_library",
         "@com_github_sourcegraph_log//:log",
         "@io_opentelemetry_go_otel//attribute",
         "@io_opentelemetry_go_otel_trace//:trace",