--- a/cmd/cody-gateway/internal/httpapi/embeddings/openai_test.go
+++ b/cmd/cody-gateway/internal/httpapi/embeddings/openai_test.go
@@ -4,9 +4,8 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
+	"math/rand"
 	"net/http"
-	"os"
 	"strconv"
 	"testing"
 
@@ -38,7 +37,7 @@ var inputSizes = []int{1, 512}
 
 func Benchmark_JsonParsing_Response(b *testing.B) {
 	for _, inputSize := range inputSizes {
-		dat, err := os.ReadFile(fmt.Sprintf("testdata/%d.json", inputSize))
+		dat, err := generateEmbeddingResponse(inputSize)
 		if err != nil {
 			b.Fatal(err)
 		}
@@ -77,7 +76,7 @@ func Benchmark_JsonParsing_Response(b *testing.B) {
 }
 
 func Test_JSON_V2_Matches_Stdlib(t *testing.T) {
-	dat, err := os.ReadFile(fmt.Sprintf("testdata/512.json"))
+	dat, err := generateEmbeddingResponse(512)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -89,4 +88,26 @@ func Test_JSON_V2_Matches_Stdlib(t *testing.T) {
 	assert.Equal(t, v1, v2)
 	assert.Equal(t, 512, len(v1.Data))
 	assert.Equal(t, 512, len(v2.Data))
+}
+
+func generateEmbeddingResponse(batchSize int) ([]byte, error) {
+	res := &openaiEmbeddingsResponse{
+		Data:  make([]openaiEmbeddingsData, batchSize),
+		Model: "text-embedding-ada-002",
+		Usage: openaiEmbeddingsUsage{
+			PromptTokens: 13,
+			TotalTokens:  42,
+		},
+	}
+	for i := 0; i < batchSize; i++ {
+		data := make([]float32, 1536)
+		for f := 0; f < 1536; f++ {
+			data[f] = rand.Float32()*2.0 - 1.0
+		}
+		res.Data[i] = openaiEmbeddingsData{
+			Embedding: data,
+			Index:     i,
+		}
+	}
+	return json.Marshal(res)
 }