… cmd/cody-gateway/internal/httpapi/embeddings/testdata/1.json
… go.sum