This switches the json parser we use for embeddings to https://github.com/go-json-experiment/json, offering >2x improvement in parsing speed (100ms improvement for a single batch!) and large reduction of memory allocations. 

We currently have 2 instances of Cody Gateway running at 50% CPU usage, and Cody Gateway overhead for this endpoint is pretty high [(p50 100ms per call, p99 500ms+)](https://console.cloud.google.com/monitoring/metrics-explorer;duration=PT5M?pageState=%7B%22xyChart%22:%7B%22constantLines%22:%5B%5D,%22dataSets%22:%5B%7B%22plotType%22:%22LINE%22,%22targetAxis%22:%22Y1%22,%22timeSeriesFilter%22:%7B%22aggregations%22:%5B%7B%22crossSeriesReducer%22:%22REDUCE_PERCENTILE_99%22,%22groupByFields%22:%5B%5D,%22perSeriesAligner%22:%22ALIGN_DELTA%22%7D%5D,%22apiSource%22:%22DEFAULT_CLOUD%22,%22crossSeriesReducer%22:%22REDUCE_PERCENTILE_99%22,%22filter%22:%22metric.type%3D%5C%22workload.googleapis.com%2Fcody-gateway.latency_overhead%5C%22%20resource.type%3D%5C%22generic_task%5C%22%20metric.label.%5C%22feature%5C%22%3D%5C%22embeddings%5C%22%22,%22groupByFields%22:%5B%5D,%22minAlignmentPeriod%22:%2260s%22,%22perSeriesAligner%22:%22ALIGN_DELTA%22%7D%7D%5D,%22options%22:%7B%22mode%22:%22COLOR%22%7D,%22y1Axis%22:%7B%22label%22:%22%22,%22scale%22:%22LINEAR%22%7D%7D%7D&project=cody-gateway-prod
), so this seems worthwhile to try.

```
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/sourcegraph/cmd/cody-gateway/internal/httpapi/embeddings
                                   │   std.txt    │               v2.txt                │
                                   │    sec/op    │   sec/op     vs base                │
_JsonParsing_Response/1/bench-10      335.4µ ± 4%   131.9µ ± 2%  -60.67% (p=0.000 n=10)
_JsonParsing_Response/512/bench-10   161.40m ± 2%   64.90m ± 1%  -59.79% (p=0.000 n=10)
geomean                               7.358m        2.926m       -60.23%

                                   │    std.txt    │                v2.txt                │
                                   │     B/op      │     B/op      vs base                │
_JsonParsing_Response/1/bench-10     171.29Ki ± 0%   20.72Ki ± 0%  -87.90% (p=0.000 n=10)
_JsonParsing_Response/512/bench-10    86.34Mi ± 0%   10.34Mi ± 0%  -88.02% (p=0.000 n=10)
geomean                               3.800Mi        468.4Ki       -87.96%

                                   │    std.txt    │               v2.txt                │
                                   │   allocs/op   │  allocs/op   vs base                │
_JsonParsing_Response/1/bench-10      1569.00 ± 0%    14.00 ± 0%  -99.11% (p=0.000 n=10)
_JsonParsing_Response/512/bench-10   793.119k ± 0%   6.156k ± 0%  -99.22% (p=0.000 n=10)
geomean                                35.28k         293.6       -99.17%
```

Closes https://github.com/sourcegraph/sourcegraph/issues/54766

## Test plan

- adds a test for comparing v2 output to stdlib
- adds a benchmark