@rafax 2024-02-13
> Pretty big file to commit - could we generate the large input instead, e.g. iterate over `n` and write `"randomkey": "randomvalue"` for each?

Fair point, switched to generating the input.